Updated tests to properly reflect desired outcome
Why: Previously the tests were testing whether or not the new
insert() mechanism had the same outcome when doing other methods.
That is actually not the desired API of insert. Insert should
be the exact same as if the data was included in the first place
and so it should check for that.