Add support for insert(elem) (#18)
* Add support for insert(elem)

Why: The ability to insert into this structure can be very important if
you have a read-heavy but right-necessary workload where rebuilding
the whole data structure would take too long

* Updated code to follow formatting standards

Mostly fixed errors with newlines. This change also edited the doc
comment for insert() to be more clear.

* Updated tests to properly reflect desired outcome

Why: Previously the tests were testing whether or not the new
insert() mechanism had the same outcome when doing other methods.
That is actually not the desired API of insert. Insert should
be the exact same as if the data was included in the first place
and so it should check for that.