fix: improve perf of count with smarter handling of half-open range
The BITS paper uses fully inclusive ranges. When this was written
I'm not sure I understood that. To get things to match the naive
version of count (find -> count) I had a while loop in the bits
count method to advance the cursor past the matched start/stop
index.

This change was found while porting mojo-lapper and removes the
while loop. It also has a more effecient branchless binary search.

Benchmarks for count improve 24-30% (see PR).